pyproject.toml
[project]
name = "epicnotify"
version = "1.0.0"
description = "Epic Games Store Free Games Notifier Bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cachebox>=5.1.0",
    "peewee>=3.18.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "simplevk>=1.0.1",
    "vk-api>=11.10.0",
]
